third_party/blink/web_tests/fast/dom/HTMLImageElement/image-picture-source-dynamic-changes-expected.txt

PASS document.getElementById("simple").clientWidth is 400
PASS currentSrcFileName("simple") is "image-set-1x.png?1"
PASS document.getElementById("simple").clientWidth is 1600
PASS currentSrcFileName("simple") is "image-set-4x.png"
PASS document.getElementById("simple").clientWidth is 75
PASS currentSrcFileName("simple") is "blue_rect.jpg?1"
PASS document.getElementById("simple").clientWidth is 400
PASS currentSrcFileName("simple") is "image-set-1x.png?2"
PASS currentSrcFileName("simple") is "image-set-1x.png?2"
PASS successfullyParsed is true

TEST COMPLETE
